==> notebooks/Spiketrain Props/04 - Compare Neuron Types.ipynb <==
from ssri_interactions.io import load_derived_generic
from ssri_interactions.stats import prop_of_total

import pandas as pd
from scipy.stats import chi2_contingency
from statsmodels.stats.contingency_tables import Table


neuron_types_all = load_derived_generic("neuron_types.csv")
neuron_types = load_derived_generic("neuron_types.csv").query("group != 'DIS'")

# is there an association between the neuron type and group?



tab = pd.crosstab(neuron_types["group"], neuron_types["neuron_type"])
pct_each = tab.apply(prop_of_total, axis=1).multiply(100).round(2)
stat, p, dof, expected = chi2_contingency(tab.values) 
tab_resids = Table(tab).standardized_resids


print(f"Chi2({dof}) = {stat:.2f}\nP = {p:.2f}")
display(pct_each)
display(tab)
display(tab_resids.round(2))

tab = pd.crosstab(neuron_types["group"], neuron_types["neuron_type"])
pct_each = tab.apply(prop_of_total, axis=1).multiply(100).round(2)
stat, p, dof, expected = chi2_contingency(tab.values) 
tab_resids = Table(tab).standardized_resids


print(f"Chi2({dof}) = {stat:.2f}\nP = {p:.2f}")
display(pct_each)
display(tab)
display(tab_resids.round(2))
